# cms/server/src/middleware/__generated__/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RoleType": {
				"enum": [
					"admin",
					"user"
				],
				"type": "string"
			},
			"IUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"programme": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/RoleType"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"first_name",
					"last_name",
					"email",
					"password",
					"country",
					"role",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedUserResponse": {
				"properties": {
					"users": {
						"items": {
							"$ref": "#/components/schemas/IUser"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"users",
					"total",
					"page",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserGetResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUser"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserPostResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUser"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.first_name-or-last_name-or-email-or-password-or-country-or-programme_": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"programme": {
						"type": "string"
					}
				},
				"required": [
					"first_name",
					"last_name",
					"email",
					"password",
					"country"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCreationParams": {
				"$ref": "#/components/schemas/Pick_IUser.first_name-or-last_name-or-email-or-password-or-country-or-programme_"
			},
			"UserUpdateResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUser"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.password-or-country-or-programme_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"programme": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"country"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserUpdateParams": {
				"$ref": "#/components/schemas/Pick_IUser.password-or-country-or-programme_"
			},
			"IModule": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"subsectionIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"createdAt",
					"subsectionIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ModulesGetResponse": {
				"properties": {
					"modules": {
						"items": {
							"$ref": "#/components/schemas/IModule"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"modules",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ModuleGetResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"subsectionIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"createdAt",
					"subsectionIds"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Page Users fetched",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedUserResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "User created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserPostResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "User fetched",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserGetResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "User updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserUpdateResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"202": {
						"description": "User deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/modules": {
			"get": {
				"operationId": "GetModules",
				"responses": {
					"200": {
						"description": "Modules Fetched",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ModulesGetResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/modules/{moduleId}": {
			"get": {
				"operationId": "GetModule",
				"responses": {
					"200": {
						"description": "Module fetched",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ModuleGetResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "moduleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}